Add "custom" option to GPIO bank name
So now the GPIO yellow block can be used to connect
to arbitrary pins listed in a platform's yaml file,
without having to mess with the GPIO mask to change
the drop down options.

This was implemented so that SNAP can use "rpi_header"
GPIOs to control the HERA phase switches.