Wishbone arbiter rearchitecture
This changes the way that the wishbone arbiter works, it should work for all
boards that currently use the wishbone bus. It dynamically generates the
arbiter in python by cascading the current wbs_arbiter. This allows for lower
fanouts on the FPGA and eases timing. Each arbiter is limited to 32 devices
before a new arbiter is added.

I also plan to remove the combinatorial logic in the arbiter which also hampers
timing closure.