Change default ports of snapshot
A previous commit had added new optional ports to the snapshot block
and had defaulted them to on.
This breaks update_casper_blocks, because the ports didn't used to exist
but will default to on, because the old block doesn't have the parameters
to turn them off.

This commit defaults to disabling the new optional ports.