Fix for issue #47
shared_bram simulation model output went through a demunge,
but the input was upstream of the munge. Now the simulation is within
both the munge and demunge.

This commit also modifies the shared bram mask, so that when bram
simulation initial values don't match the size of the underlying RAM
they value vector will be truncated or zero padded to match. A warning is
raised. This prevents the simulation from erroring out.