Modified White Gausian Noise and u2n blocks
In the "Sources" section of the Casper DSP library, the recent merge with
casper-astro library broke our dsim, inside is a log-lut which pre-merge
had F_11_9 output, and post-merge had F_9_7.

I have parameterised this bitwidth and binpoint, defaulting to F_9_7, so as
not to break whoever else's design when they merge and update their blocks,
but I have updated our dsim design to specify F_11_9, so it should work
in subsequent updates and merges too.