Create CODE_OF_CONDUCT.md
As suggested at the CASPER 2018 meeting, we should have a code of conduct. This one is based on the 'Citizen code of Conduct', and is more appropriate than the other Github templated 'Convenant code of conduct' as it covers meetings. Could also consider basing on http://www.astropy.org/code_of_conduct.html