Fix 'xps_build_new_library' bug in new Matlab ver.
The adc083000x2 yellow block included a 'conv' library block that
caused issues with the 'xps_build_new_library' script in versions
of Matlab newer than R2016b. Considering the adc083000 is not
supported with the SNAP or SKARAB currently, the 'conv' block was
removed.

If 'adc083000' yellow block support is added in the future, note
that the data streams now expect HDL code with 2's complement.