Move +1 in rx bytes into gbe_rx
The code was arranged so that the reported number of bytes
received by the CPU was one less than the actual number available.
The gbe_cpu_attach_wb module dealt with this.

This commit refactors so that the reported number of bytes available
from gbe_rx reflects the actual number of bytes available. This seems
more reasonable.