Add #defines for enabling XADC / SPI flash / ICAP
This allows the same codebase to be used with the VCU118, which
doesn't support these things (yet).

Probably need to think this through a little more, as the number
of supported platforms begins to spiral out of control.

Currently to compile for VCU118 no extra defines are needed.
To compile for SNAP define USE_XADC, USE_SPI, and USE_ICAP