Correct core_info tab generation for AXI4-lite.
Did some gross stuff in axi4lite_memory_map() in order to pass the write_core_info functions
the list of axi4lite_devices in the 'proper' format for it to make core_info.tab as desired.

Passed fullpath to axi4lite_interconnect yellow block for it to know the design name.