Red pitaya bram support
1. Added support for the BRAMs in the RP flow
2. Fixed up the clocking and reset architecture.

BRAMs not working 100% yet as there is some issue with the AXI2VHDL
generation that wont allow me to put the correct address space or
maybe it expects it in bytes not words, I just dont care anymore!
And I want to go home. Why oh why, Adamm did you have to volunteer
us for the hardware porting workshop?