Order axi memory devices largest to smallest
To try and comply with an xml2vhdl requirement that an
N-byte device be aligned on an N-byte boundary.
If all devices are 2^m bytes, this should work.

It's unclear to me whether the alignment needs to be
at the axi device level, or the sub-register level.
If the former, then I don't think this will fix having (eg)
random numbers of registers, since the total size of the sw_reg
space will not be a power of 2 bytes (even though each register
within the space is 2^2 = 4 bytes in size)