More AXI generation fixes
Had an issue where, if there were (eg) three 4-byte registers
on the sw_reg device bus, the xml2vhdl code would generate
a device with `byte_size="8"`.
Traced this to an (I think) erroneous calculation of size in
the xml2vhdl package.

This commit also tweaks the verilog.py source so that sw_reg
is not always placed last, but is placed in accordance with its size.
This fixes my misunderstanding -- the top-level devices need to be
N-byte aligned, not the underlying (32-bit) registers. Therefore
we can't assume that the sw_reg block should go last. It may be
relatively large.

Also round up memory spaces to 2**n bytes before ordering. This
shouldn't change anything, but means the calculation done by the toolflow
and xml2vhdl should match.