Fix issue #46
Give the software register (FPGA -> CPU) one simulation output
for each of the input signals.
Because I'm lazy, and don't want to mess around with slicing/dicing/casting
the output, just hook up simulation gateways directly to the block inputs.
Match the latency of the real pipeline, and assume (optimistically?) that the
compiler will be smart enough to optimize away the non-HDL gateways