This fixes a bug we found in the fortygbe counters
The coutners were under counting the number of packets send.
Turns out that the signals driving the enable ports on the counters were
coming from a different clock domain causing the counters to miss some enable
signals. Interesting bug and a little tricky to find.

The counters now align with the packet counters on the switch.