Implement some windowing functions and sinc with python
This avoids the need for the DSP toolbox when generating PFBs.

I followed the process in https://uk.mathworks.com/help/matlab/matlab_external/call-user-defined-custom-module.html
to implement these functions in user python.

Note that it would be easier/better to just use the numpy versions of the various
windows it supports. But this crashes MATLAB as per
https://uk.mathworks.com/matlabcentral/answers/503505-why-matlab-crashes-when-i-try-to-use-a-numpy-python-package-2019a

So for now, only hamming / hann / blackman are implemented