Skarab 40gbe changes: Fixed wb device order bug
The wishbone devices which requested an address would change order from one
compile to the next. This meant that the order that the devices are passed to
the arbiter is not sequential and the arbiter breaks. Causing devices to be
unreadable. I have reordered the list of wb_devices according to address prior
to generating the arbiter and this solves the problem.

This commit also includes changes to the peralex registers to include a present
signal from each of the ethernet cores. This allows the microblaze to tell
which cores are present and only initialise those ones.

There is also a new microblaze which supports the dynamic 40gbe core changes.