"Fix" flash read ID issue
In compiles from Vivado 2019.1.2 (and possibly earlier,
though the issue was not present in 2016.4) reading the flash ID
failed, and would return all F's. Preceding the ID read with
some other flash_read call, or reading the ID twice, would work.

This issue doesn't seem to be cured by:
  - Updating the bsp to SDK 2019.1
  - using a known good microblaze binary generated in 2016.4

This commit opts to simply read the ID twice. Empirically, this
seems to work fine. It doesn't fix the underlying issue, nor
is it guaranteed that if some other flash transaction becomes the
first to be issued by the microblaze this won't fail too.