Fix missing core_info final entry
Fix issue where the memory file is divided into 32-byte
lines, and any leftover contents is discarded.
This commit rounds up the number of lines to make sure nothing gets missed.
It's quite amazing that this managed to not cause issues in the past --
the effect is that when using JAM, there is a high-probability the
alphabetically last software core will be inaccessible by name.