Come out of 4B address mode after every access
I may have got over-excited before.

The Xilinx bitstream has to be compiled for either 3 or 4 byte
address mode. If firmware changes the behaviour of the chip then
the golden image will not boot.

Probably the "best" thing to do would be to set all the flash chips
to permanently use 4B addressing in a non-volatile way and then
generate appropriately configured bitstreams. But a)
that is not their factory default behavior, and b) this
would lead to catastrophic backwards compatibility issues with
boards using an old 3B-mode golden image.

Basically, I got ideas above my station, and shouldn't have messed
with LS's code.