Make software-controlled reorder map have 32-bit wide bram
This is probably wasteful for many reorders. *But* it makes
the control software (which usually forces 32-bit aligned read/writes)
easier. "Minimum RAM" vs "easier software" could probably
be a parameter, but just doing the easy thing is likely to be
what the user wants almost all of the time