Make YellowBlock aware of block type and use for port naming
PR demands that the interface to a reconfigurable module doesn't
change between module versions. This is hard to do at the best of
times, but particularly difficult when port names depend on simulink
model name.

This commit adds a `blocktype` attribute to the YellowBlock class,
which holds the name of the block class (i.e., the block tag).

The 100G block yellow block has been changed to generate top-level
ports which depend on this class name and the port number, not
on the model name.

The DTS module is also changed to have a static top-level port prefix.