Fix for the skarab 40gbe port swapping issue
We noticed that when using more than 1 40gbe port on a skarab the ports would
be swapped. ie control packets from the microblaze would come out on the incorrect
interface. This was traced down to a miss-mapping of the wd_ids in top.v to the
list of hig and los addresses in the wb_master_arbiter.v This was caused by the
changes to make the toolflow deterministic by sorting the list of wb_devices.
It turns out that the wb_ids were generated prior to the sorting and this caused
the problem.

But all is well now in casperland and we can all move on with our lives.