Remove `design_name` from internal AXI signals
Why does it even need to be there at all? The design name doesn't
add any uniqueness to any signal names -- it's the same for all yellow blocks.
Either this entire field could be removed altogether, or, it really needs
to be unique. Replace it with the register name, which should be unique
(blocks are obliged to ensure they generate register names which don't clash
with one another).