Final sw_reg int/string default_val fix
"To Processor" software registers should have no default value
set. This wouldn't make sense anyway, since these registers are
overwritten by the fabric every clock.

Previously, software registers of both directions were given a default
value, and if - for a To Processor register - the value was zero, the toolflow
treated it as a register which should be written to from the fabric.
Previous commits make 0 an allowable default, and declare default_val=None to
be the indicator that a register should be fabric-writable.

This commit is [hopefully] the last fix in this series.