Make source port = dest port.
This is silly, but currently there isn't a simulink input of src_port.

In the other GbE blocks, source port is the port the Ethernet core listens on
(I think). This signal is easily available in HDL, but is on the wrong clock domain,
so currently I'm not using it.