Add "Idle CS" user-defined idle state for the CS line
For true SPI, when no transaction is occuring, CS should
be in it's inactive state.
For some systems (eg the SNAP2 ADS5296 ADC) the CS lines actually
drive a GPIO -> CS converter, so a different idle state may be required.

This commit make the idle state a port on the SPI block,
and adds control of this port via wishbone