Add v2 of the ADS5296 interface.
As currently configured, this uses a deserialize-by-4 to lower
the clock rate and then uses the (assumed synchronous) fclk
signal to figure out word boundaries as the words are built back up
to 10 bits. This is all to work around the fact that arbitrary bit-width
ISERDES blocks aren't supported in Ultrascale.