Initial commit of rfdc yb
So far this is able to get through the toolflow to generate a
bitstream (albeit a simple one tile, one adc, with PLL) and has not yet been
tested on a platform. More testing needs to be done.

This rfdc yb strives to make available the custom flexibility of the rfsoc. This
support is achieved by mapping simulink mask variables to vivado parameters and
using vivado tcl commands to configure the IP core. Currently, as supported by
the simulink mask and subsequently the configuration for either a dual tile or
quad tile adc is replicated across all enabled tiles.

Future support could either extend the capability to the user to have different
adc configuration across tiles or other additional capabilities can be done to
limit the simulink mask to either load a default configuration or have a
platform default configuration and then enable the current ability to configure
the adcs as an 'advanced user configuration' mode.

To facilitate the mapping from simulink mask variables to vivado parameters some
dummy class containers are used to represent rfdc `tiles` and `adcs`. These
containers are populated from the simulink mask block and then mapped to vivado
parameters using the rfdc class `attr_map` attributes to resolve the
corresponding vivado param. Additionally, `value_map`s are used to map simulink
mask values (represented as strings) to vivado param values (typically
enumerated values, floats or boolean).