Improve 1GbE constraints
Don't include clock constraint for GTX-based 1GbE, since in this
case the IP declares the clock and attempting to declare it
again throws a critical warning.

Make the wire sys_clk have a keep=true attribute, so that it can
be used for clock declarations. Previously the pin (eg "sys_clk_p")
and/or clock names (eg "sys_clk_p_CLK") were used, which are
inconsistent across platforms.