Beginning of incorporating board designs as the way to work with IP
Board designs are becomming mroe and more required to support the full suite of
xilinx chip capabilities. It seems to make more sense then that when intancing a
platform such as the zcu216 (or any rfsoc platform) that a base board design is
used and the yellow blocks will expect that his block design be inside the base
board design or at least insert itself into the design.

The board design is necessary because at is the only way that a hardware
platform can reliably be created. There are a lot of methods/work arounds that
one could do to update or modify your hardware platform (primarily by writing
and maintaining your own device tree overlays) however, it makes more sense to
just leverage the xilinx flow as they already have done the work to provide you
with those outputs.

All of this is of primary interest for the RFDC. Because in order to support the
xrfdc driver the device tree node must be specified and part of the drivier
initialization requires the param-list of the device node to be generated, not a
trivial task. The xrfdc driver is the only practical way to initaite multi-tile
synchronization and users will therefore want to be able to use this capability.