working snapshot example design using all four tiles
Updates `toolflow.py` to create project in `init` stage and then the platform
yellow block (e.g., zcu216) has an `init` call that instances the board design.
This allows allows other yellow blocks to now access the platforms board design
and either add IP to the board design or update IP already placed in a sourced
board design. This seemed to be the quickest and least painful of all the
methods I looked at in building out and supporting zynq, the rfsoc and
downstream software flows as this allows for the ability to get the xilinx
shell archive (`.xsa`) to include IP like the MPSoC and rfdc for the full
supported xilinx software flow.

That being said, if using board design is not really where we want to go than
there are still things we can do to support instancing IP from the IP catalog
instead of a board design. It would just require our own hacking of device tree
generator stuff which I have demonstrated for the rfdc.

Will go back now to clean up code but needed to at least capture a working
design to go back to. Tested with casperfpga in py3.

No particular order of things still needing to do and by no means exhaustive,
just a brain dump...

 * rfdc yellow block
    - fix up any unique name port connection issues
    - implement mts and test adding related IP needed for that, this may require
      doing something in the rfsoc platform block (e.g., zcu216) to provide the
      clocking infrastructure to support and so there could be this back and for
      that needs to happen with the dependencies. At first glance the zcu216
      will need to provide the required clock that the rfdc can do a fabric copy
      for the mts/mcs algorithm
    - give board design instance a unique name
    - document/comment to make more transparent the board design stuf
    - rfdc axis reset is connected to axil reset, worth re-thinking, was just to
      make simple at the time, not sure what most other casper adcs do

 * rfdc simulink
    - still need to extend to dual-tile parts, but eveyrthing should work break
      well with any quad part
    - port placement disabling and enabling ports is wonky, better port drawing
      neededed

 * zcu216 yellow block
    - need better handling of clocking infrastructure and pass of of information
      between rfdc. Problem being that the lmk/lmx clocking and user clock
      selection in the platform block dependent on rfdc configuration
    - implement mts clocking support (see rfdc yellow block mts note)
    - I still have a goal to become board design agnostic... but this is a good
      first pass... will want to go back to this commit and previous to see what
      I was doing to get board design agnostic potentially leverage more board
      design tcl calls like rfdc instead of adding direct from ip catalog
      doesn't need to be that way for custom hdl though only xilinx specific and
      things we care about to be in software design (e.g., mpsoc, rfdc)