Add support and simple snapshot test for the htg zrf16 49dr board
The `test_htg` model is able to build out for the 49dr part. But fails for the
29dr version. This is most liekly due to the lack of implementing out support
for the gen 1 rfsocs in the xps library mask and rfdc yellow block. Need to
investigate. Will go back and extend/improve when also adding support for the
rfsocs on the zcu208/zcu111 as they are dual tile architectures that also needs
to be supported in addition to any gen 1 quirks in the xps library mask.

The zrf16 made by htg has two versions of the board. One with a 1st gen 29dr
rfsoc and the other that is a 3rd gen 49dr rfsoc. The boards I have worked on
are almost identical in layout except for the PS DDR4 configuration on the 49dr
is different.

The approach here is then to have a single zrf16 platform to instance that
selects between the two revisions of the board. This seemd to make the most
sense as to try to achieve a way to support to boards without maintaining two
separate yellow blocks.

I had also explored how to get a single platform configuration file (.yaml) with
having a list of fpga part numbers in the 'fpga' dict value and have the
toolflow determine the correct one and update the platform dictionary based on
the `hw_sys` selected  but that proved to be a little to difficult at the moment
with how the platform/xsg blocks work together and instance each other. It turns
out the xsg block runs first preparing the targeted platform yellow block. Then
all other yellow blocks run.

Even if the xsg yellow block was edited to determine the correct part based on
hw_sys the toolflow is passing the platform dictonary to the yellow blocks and
the xsg version that updated `fpga` key in the platform dict will get passed to
the final platform (e.g., zrf16) but not to the other yellow blocks as a copy is
not sent back up.

But, here a single `zrf16` yellow block is maintained by a factory method in the
`zrf16_29/49dr` yellow blocks that end up instancing the common `zrf16`
implementation.