Include Simulink/SysGen design as IP, not HDL netlist
Change sysgen build script so that design is compiled to Vivado IP.
Change jasper.per block generation to add an extra IP yellow_block
to reference this.

This is
a) neater -- this is how sysgen designs are meant to be
included in Vivado projects.
b) Tolerant of things like HDL files being part of VHDL libraries,
which the toolflow doesn't like because it imports all HDL files
into the default lib