Start of making rfdc independent of the xilinx xsa
The rfdc is still added to the block design but instead of enabling a seperate
AXI interface for interfacing with the rfdc AXI memory map it will still be
controlled by the CASPER AXI4-Lite interconnect block. This complicates a little
bit the expected software flow for the rfdc as the xilinx block design is used
to create the `.xsa` that will populate driver information for all slaves with a
`memory path` to the MPSoC. Since we are just adding the rfdc as a standalone
block with no connection to the MPSoC this path does not exist and therefore
the rfdc driver will not automatically be registered in the software design from
the `.xsa`.

All is not lost though. With the rfdc still in the block design its IP
configuration is still present in the hardware design. It just won't have a
software design component. So instead, we can still use the hardware design
information to manually create our own software output produts. Here we are only
really concerned with using the already builtin xilinx rfdc driver. So we only
need the device tree overlay as required by that driver. This process would be
the same though if CASPER were to wish to create more sophisticated drivers for
our IP (e.g., when we say port the tools to vitis this is what we would do).

This commit stages the new block designs and rfdc yellow block for this change.
The software part of creating the rfdc device tree overlay has been done. It
just needs to be wrapped into the tools (those files not added here yet).

Additionally, modified the `gen_xps_add_design_info` to include the rfdc in the
meta data infromation for use in caperfpga.

Misc. changes to rfdc for extending the NCO are also added here (just hard coded
and commented out though as I needed them to just test something in the lab --
need to extend to the simulink mask).