Downgrade error to warning
SKARAB might require QSFP ports to be used in sequence, but generally
this isn't the case.

TODO: Either check the platform type in simulink, or do so in the yellow
block python and apply check as appropriate.

This commit just downgrades the error to a warning. The 40G yellow block
is almost certainly broken for SKARAB in this mlib branch anyway(!)