As good as its gonna get last minute before casper workshop
* Finish adding and testing having all rfdc ports in the block design external
and then using jasper to wire it all together. This gives us the advantage of
having the IP in the block design. We don't get everything we want with by
being able to use the full software project flow of hsi to generate the device
tree overlay. This is because vivado/hsi/vitis cannot resolve a clear memory
peripeheral path. There seems no way that I could see to declare that path by
declaring in vivado's memory spcae manager a specifc address space for the rfdc
within the `M_AXI` that was being used to connecet to the casper axi4-lite
interconnect. But we can export the hardware definition and the rfdc IP is
declared within the hardware project brining everything along with it so we can
use the `hsi` calls to build the device tree overlay ourselves and augment what
`hsi` outputs. This will be the next step, to automate this flow. But the main
components are in `hsi-plnx`.

* added pynq2x2 mask, platform definition, and yb. The yb is complete (as it is
similar to the zcu111) but the platform definition needs to be finished in order
to build out. Not going to be ready by conference as spent most time working on
tcpborphserver the last couple weeks.

Other last minute changes. Very likely more bugs introduced in doing this but at
the very least have tested rebuilding examples against the zcu111/zcu216/zrf16
and still works:
  * fix adc `clk_out` and pll `ref_clk` lists not updating in gui
  * fixed problem with disabled tiles still being editable
  * fixed not being able to get 8 samples w/ 1x dec at max fs on zcu111
  * add fine and coarse nco frequency options (simulink and rfdc yb)
  * add analog control settings - Nyquist zone/Calibration mode (simulink and yb)
  * update xps_library with rfdc changes to analog/nco freq controls
  * update xps_library to at least show the pynq2x2