intermediate updates to hsi/xsct scripts for rfdc dt
most likely will remove these from the repo to implement
within the yellow block. Until much more of the Vitis/software/device
tree support is integrated it would make more sense for everything to
be local to within the rfdc. When more functionality is added for xsct
support it could be more advantages to have more scripts or structure