Update rfsoc platform support for dt generation with .xsa export w/ bitstream
Also more formally support the pynq rfsoc2x2 by finishing out its platform
definition, yellow block creatino and adding to the xps library

zrf16_29dr base board design updated to use the correct ddr for the board. It
had been working but noticed that the ddr was different than the original
schematic which is what the board design was set for.