experimental addition of full xsct/dt creation in toolflow
This somewhat marks a milestone in supporting rfsoc/rfdc within the toolflow.
Because with these changes here in addition with supporting software from
casperfpga/tcpborphserver a design can be created from simulink and loaded with
casperfpga and functionality of the rfdc and board can be done without any other
manual bootstrapping.

This current commit will break for non zynq soc's. It merley is an intermediate
staging of complete and successful integration of the rfdc from platform, to yellow block,
to the linux software driver. The hsi/xsct/py scripts that were added have been
removed in favor of just tracking the required within the rfdc yellow block.
This was a choice made to just keep the required dt generation within the rfdc.
Because at this point, while any MPSoC with a board desgin can be used to create
a valid Vitis platform and software project we are really only using it within
the context of rfdc and rfsoc. However, the way it is implemented *does not*
limit the ability to extend further. Instead, it just provides a convenient way
to keep functionality where it needed most until (or maybe it never happens)
when more of this functionality wants to be extended. At that point perhaps a
more general implementation.

From here is to start a small Vitis back end class that will allow compatability
with all of the tools