Fixing sw_reg issue
The to processor sw_reg on the axi bus would not be connected up to the
sysgen block correctly without the this change to the port names. Not
sure how it worked previously.

Fixed issue in toolflow.py that occured after the merge, where the
add_compile_cmds function requries a parameter threads.