Add ability to replace CMAC after placement
Effectively, this commit adds support to the "new" way of dealing with
100G IP (i.e., creating one piece of IP per core) so that the "old"
way is still supported (using IP and then unplacing and replacing the
CMAC).

This is (seems to be?) necessary for the ZRF16 board, which has the
GTY transceivers laid out such that IP _can't_ be generated for both
each 100G core, because one of them must use a CMAC from a different
SLR (which the IP configurator forbids).