working version for the rfsoc4x2, but breaks any other board
there are still assumptions that need to be made regarding the connections and
how to wire them up. This needs to be worked out next.

Also the approach to adding a block design to the `VivadoBackend` class requires
that it doesn't break the flow for designs that do not use a board design. This
needs to be tested.