First pass working version using vivado block design
Here, the creation of a vivado block design inserts itself as a
`create_bd` stage. This is done during the `add_compile_cmds` of the backend
with each yellow block implementing a `modify_bd` method.

Considering that the generation of a block design is a list of tcl cmds it would
be possible to have this functionality be contained within the `gen_tcl_cmds`.
However, the idea here is to provide an infrastructure with the block design
(.bd) file treated more as a source like the top module being a Verilog class.
To do keep this in `gen_tcl_cmds` as it is implemented now would require a
modification to that structure, or to have the block design more amorphous to
the process.

However, considering this, and that the rfdc still uses `gen_tcl_cmds` that is
accomodated here by waiting until all yellow blocks have a chance to modify
block design and using the save/validate as part of a pre-synth stage. The idea
is to migrate the rfdc to use a `modify_bd` unless otherwise reverting back to
`gen_tcl_cmds` for everything.