removed bus_expand_callback
The bahaviour of mask callbacks seems to have changed in Simulink.

When setting up an fft_wideband_real, the first change of parameters
succeeded but sebsequent changes caused an error while the init
scripts ran.

The mask callback on some parameters in the bus_expand block
caused a cascade of calls to itself, as well as bus_expand_init that
resulted in an erroneous error.

The solution has been to remove the mask callback. This means that
the mask does not change depending on parameter changes as before
but this does not change operation, it was a nicety for the user.