suppress mask initialization output from datatype logic
compiling with these blocks prior to this change would result unspressed output
to the command prompt such as `f_type = 'fixed'`, `f_type = 'custom'`,
`enfixpt = 'on'`, and others.

This was because mask initializations in the `cmult`, `butterfly`, `biplex`,
`fft_unscrambler` for data type mask parameters executed and variables used to
switch configuration based on type was unsuprssed (not ending in `;`).