Fix simulation error with swreg in m2021a
Changed data type rule to output of gain block ot also be 32-bit

The simulink gain block is for left shifts. This is needed for example when
spliting a 32-bit swreg into multiple bitfields. The default rule for the gain
block with fixed point data types is for full precision and so the output type
is at least twice the input type. This seems to cause errors with this version
of  sysgen/model composer as now the gateways are enforcing checks that input
bitwidths are compatible.