Async restore, more intelligent fanout control
Improved the fanout delay calculation
1. If the maximum fanout allowed is 1 or less, then have a minumum of 1
delay to control fanout, otherwise allow it to go to 0. So if we are
comfortable with fanout, then allow resource reduction to 0.
2. The fanout calculation is now more accurate.

The default is that DUMB_BRAMS is now 0. The logic is still there though
for those who prefer this setting.

This may mean that designs that previously made timing may not, and
changes to the maximum fanout parameter will be needed.