bugfix: PFB FIR coefficients not symmetric
The window coefficients for the PFB FIR were not quite symmetric and
were off by half.

This was previously thought to be intentional, and extra logic was
required to allow only half the coefficients to be
stored, but for them to be output asymmetrically.

This has been fixed so that coefficients are now symmetric. This
required removal of this extra logic (that should improve timing
andreduce resource use), and a change in the generation logic so that
the coefficients are symmetric (this should improve PFB channel response
in a small way).

This change should be backwards compatible i.e no changes to existing
designs should be required.