split buffer implementation into taps and coeffs
Data taps and coefficient storage implementation are now separate
options.
Distributed memory removed as an option as scripts can't draw these
properly as the ports change. (This could be added with a little effort
but would only make sense for small PFBs).
Removed the checks of the implementation target at a high level as
checks already done at a low level.