Added 'Ultra RAM' option to top level parameters of 'delay_wideband_prog' block.
Known Issue: Memory has to be 'dual port' for BRAM and Ultra RAM selection. Chnaging to 'single port' generates an error.
Also - no support for 'Distributed memory' added - this would require modification of other scripts. If 'Distributed memory'
is required, it would be necessary to comment out the init script and manually set the memory type in the block.