Changes to support multicast and farm mode in the alveo 100gbe
Farm mode is untested and requires the bram for the LUT to be prepopulated
at compile time, there is no support for this from the toolflow, so a work-
around will need to be implmented.

The multicast is untested at this stage and will probably remain so, as we
are moving on to other work.