fixing more dac inconsistencies
This should fix more than it breaks...

Simplifies the mask logic for the dac in a few places. Continues reconciling
tile/slice architecture from previous commit with changes to
`get_rfsoc_properties` and utilize that information to simplify logic and begin
removing static hard-coded variables.

This also fixes a bug where the dac mixer mode could not be configured in coarse
operation. It allowed for it to be selected but always defaulted back to fine

There are still issues configuring outputs in I/Q mode. The change in logic has
broken some of that capability. There are also still some bugs with enable
option for odd slices not being selectable when it should be. This typically
requires toggling the even slice neighbor a few ways to get the odd slice back.